+ max table children, deprecate old visualize
