+++ openadapt/visualize.py
@@ -1,159 +1,137 @@
+from base64 import b64encode
+from functools import partial
+from os import path, sep
 from pprint import pformat
-from threading import Timer
-import html
-import os
-import string
 
-from bokeh.io import output_file, show
-from bokeh.layouts import layout, row
-from bokeh.models.widgets import Div
 from loguru import logger
+from nicegui import ui
 from tqdm import tqdm
 
-from openadapt.crud import (
-    get_latest_recording,
-)
-from openadapt.events import (
-    get_events,
-)
+from openadapt import config
+from openadapt.crud import get_latest_recording
+from openadapt.events import get_events
+from openadapt.models import Recording
 from openadapt.utils import (
+    EMPTY,
     configure_logging,
     display_event,
-    evenly_spaced,
     image2utf8,
-    EMPTY,
+    plot_performance,
     row2dict,
     rows2dicts,
 )
 
-from openadapt import config
-
 SCRUB = config.SCRUB_ENABLED
+
 if SCRUB:
+    # too many warnings from scrubbing
+    __import__("warnings").filterwarnings("ignore", category=DeprecationWarning)
     from openadapt import scrub
+
 
 LOG_LEVEL = "INFO"
 MAX_EVENTS = None
-MAX_TABLE_CHILDREN = 5
-MAX_TABLE_STR_LEN = 1024
 PROCESS_EVENTS = True
-IMG_WIDTH_PCT = 60
-CSS = string.Template(
+
+performance_plot_img: ui.interactive_image = None
+
+
+def create_tree(tree_dict: dict, max_children: str = config.MAX_TABLE_CHILDREN) -> list[dict]:
     """
-    table {
-        outline: 1px solid black;
-    }
-    table th {
-        vertical-align: top;
-    }
-    .screenshot img {
-        display: none;
-        width: ${IMG_WIDTH_PCT}vw;
-    }
-    .screenshot img:nth-child(1) {
-        display: block;
-    }
+    Recursively creates a tree from a dictionary.
 
-    .screenshot:hover img:nth-child(1) {
-        display: none;
-    }
-    .screenshot:hover img:nth-child(2) {
-        display: block;
-    }
-    .screenshot:hover img:nth-child(3) {
-        display: none;
-    }
+    Args:
+        tree_dict (dict): The dictionary to create a tree from.
+        max_children (str, optional): The maximum number of children to display. Defaults to MAX_TABLE_CHILDREN.
 
-    .screenshot:active img:nth-child(1) {
-        display: none;
-    }
-    .screenshot:active img:nth-child(2) {
-        display: none;
-    }
-    .screenshot:active img:nth-child(3) {
-        display: block;
-    }
-"""
-).substitute(
-    IMG_WIDTH_PCT=IMG_WIDTH_PCT,
-)
-
-
-def recursive_len(lst, key):
-    try:
-        _len = len(lst)
-    except TypeError:
-        return 0
-    for obj in lst:
-        try:
-            _len += recursive_len(obj.get(key), key)
-        except AttributeError:
+    Returns:
+        list[dict]: Data for a Quasar Tree.
+    """
+    tree_data = []
+    for key, value in tree_dict.items():
+        if value in EMPTY:
             continue
-    return _len
+        node = {
+            "id": str(key)
+            + f"{(': '  + str(value)) if not isinstance(value, (dict, list)) else ''}"  # dynamic f-string
+        }
+        if isinstance(value, dict):
+            node["children"] = create_tree(value)
+        elif isinstance(value, list):
+            if max_children is not None and len(value) > max_children:
+                node["children"] = create_tree(
+                    {i: v for i, v in enumerate(value[:max_children])}
+                )
+                node["children"].append({"id": "..."})
+            else:
+                node["children"] = create_tree({i: v for i, v in enumerate(value)})
+        tree_data.append(node)
+    return tree_data
 
 
-def format_key(key, value):
-    if isinstance(value, list):
-        return f"{key} ({len(value)}; {recursive_len(value, key)})"
-    else:
-        return key
+def set_tree_props(tree: ui.tree) -> None:
+    """
+    Sets properties for a UI tree based on values from config.
+
+    Args:
+      tree (ui.tree): A Quasar Tree.
+    """
+    tree._props["dense"] = config.DENSE_TREES
+    tree._props["no-transition"] = config.NO_ANIMATIONS
+    tree._props["default-expand-all"] = config.EXPAND_ALL
+    tree._props["filter"] = ""
 
 
-def indicate_missing(some, every, indicator):
-    rval = []
-    some_idx = 0
-    every_idx = 0
-    while some_idx < len(some):
-        skipped = False
-        while some[some_idx] != every[every_idx]:
-            every_idx += 1
-            skipped = True
-        if skipped:
-            rval.append(indicator)
-        rval.append(some[some_idx])
-        some_idx += 1
-        every_idx += 1
-    return rval
+def set_filter(
+    text: str,
+    window_tree: ui.tree,
+    action_tree: ui.tree,
+) -> None:
+    """
+    Sets filter for UI trees.
+
+    Args:
+        text (str): The text to filter by.
+    """
+    window_tree._props["filter"] = text
+    action_tree._props["filter"] = text
+    window_tree.update()
+    action_tree.update()
 
 
-def dict2html(obj, max_children=MAX_TABLE_CHILDREN, max_len=MAX_TABLE_STR_LEN):
-    if isinstance(obj, list):
-        children = [dict2html(value, max_children) for value in obj]
-        if max_children is not None and len(children) > max_children:
-            all_children = children
-            children = evenly_spaced(children, max_children)
-            children = indicate_missing(children, all_children, "...")
-        html_str = "\n".join(children)
-    elif isinstance(obj, dict):
-        rows_html = "\n".join(
-            [
-                f"""
-                <tr>
-                    <th>{format_key(key, value)}</th>
-                    <td>{dict2html(value, max_children)}</td>
-                </tr>
-            """
-                for key, value in obj.items()
-                if value not in EMPTY
-            ]
-        )
-        html_str = f"<table>{rows_html}</table>"
-    else:
-        html_str = html.escape(str(obj))
-        if len(html_str) > max_len:
-            n = max_len // 2
-            head = html_str[:n]
-            tail = html_str[-n:]
-            snipped = html_str[n:-n]
-            middle = f"<br/>...<i>(snipped {len(snipped):,})...</i><br/>"
-            html_str = head + middle + tail
-    return html_str
+def toggle_dark_mode(
+    ui_dark: ui.dark_mode, plots: tuple[str], curr_logo: ui.image, images: tuple[str]
+) -> None:
+    """
+    Handles dark mode toggle.
+
+    Args:
+        ui_dark (ui.dark_mode): The dark mode switch.
+        plots (tuple[str]): The performance plots.
+        curr_logo (ui.image): The current logo.
+        images (tuple[str]): The light and dark mode logos (decoded)
+    """
+    global performance_plot_img
+    ui_dark.toggle()
+    ui_dark.update()
+    curr_logo.source = images[int(ui_dark.value)]
+    curr_logo.update()
+    performance_plot_img.source = plots[int(ui_dark.value)]
+    performance_plot_img.update()
+
 
 @logger.catch
-def main():
+def main(recording: Recording = get_latest_recording()) -> None:
+    """
+    Visualize a recording.
+
+    Args:
+        recording (Recording, optional): The recording to visualize. Defaults to get_latest_recording().
+    """
     configure_logging(logger, LOG_LEVEL)
 
-    recording = get_latest_recording()
+    ui_dark = ui.dark_mode(config.VISUALIZE_DARK_MODE)
+
     if SCRUB:
         scrub.scrub_text(recording.task_description)
     logger.debug(f"{recording=}")
@@ -167,27 +145,105 @@ def main():
     logger.info(f"event_dicts=\n{pformat(event_dicts)}")
 
     recording_dict = row2dict(recording)
+
     if SCRUB:
         recording_dict = scrub.scrub_dict(recording_dict)
 
-    rows = [
-        row(
-            Div(
-                text=f"<style>{CSS}</style>",
-            ),
-        ),
-        row(
-            Div(
-                text=f"{dict2html(recording_dict)}",
-            ),
-        ),
-        row(
-            Div(
-                text=f"{dict2html(meta)}",
-                width_policy="max",
-            ),
-        ),
+    # setup tables for recording / metadata
+    recording_column = [
+        (
+            {
+                "name": key,
+                "field": key,
+                "label": key,
+                "sortable": False,
+                "required": False,
+                "align": "left",
+            }
+        )
+        for key in recording_dict.keys()
     ]
+
+    meta_col = [
+        {
+            "name": key,
+            "field": key,
+            "label": key,
+            "sortable": False,
+            "required": False,
+            "align": "left",
+        }
+        for key in meta.keys()
+    ]
+
+    plots = (
+        plot_performance(recording.timestamp, save_file=False, show=False),
+        plot_performance(
+            recording.timestamp, save_file=False, show=False, dark_mode=True
+        ),
+    )
+
+    with ui.row():
+        with ui.avatar(color="auto", size=128):
+            images = ()
+
+            # generate base64 encoded images for light and dark mode
+            for i in range(2):
+                logo_base64 = b64encode(
+                    open(
+                        f"{path.dirname(__file__)}{sep}app{sep}assets{sep}logo{'_inverted' if i > 0 else ''}.png",
+                        "rb",
+                    ).read()
+                )
+                images += (
+                    bytes(
+                        f"data:image/png;base64,{(logo_base64.decode('utf-8'))}",
+                        encoding="utf-8",
+                    ).decode("utf-8"),
+                )
+            curr_logo = ui.image(images[int(ui_dark.value)])
+        ui.switch(
+            text="Dark Mode",
+            value=ui_dark.value,
+            on_change=partial(toggle_dark_mode, ui_dark, plots, curr_logo, images),
+        )
+
+    # create splitter with recording info on left and performance plot on right
+    with ui.splitter(value=20).style("flex-wrap: nowrap;") as splitter:
+        splitter._props["limits"] = [20, 80]
+
+        # TODO: find a way to set "overflow: hidden;" for the splitter
+        with splitter.before:
+            ui.table(rows=[meta], columns=meta_col).style("min-width: 50em;")._props[
+                "grid"
+            ] = True
+        with splitter.after:
+            global performance_plot_img
+            performance_plot_img = ui.interactive_image(
+                source=plots[int(ui_dark.value)]
+            )
+            with performance_plot_img:
+                ui.button(
+                    on_click=lambda: plot_performance(
+                        recording.timestamp, show=True, save_file=False
+                    ),
+                    icon="visibility",
+                ).props("flat fab").tooltip("View")
+
+                ui.button(
+                    on_click=lambda: plot_performance(
+                        recording.timestamp, save_file=True, show=False
+                    ),
+                    icon="save",
+                ).props("flat fab").tooltip("Save as PNG")
+
+    ui.table(rows=[recording_dict], columns=recording_column)
+
+    interactive_images = []
+    action_event_trees = []
+    window_event_trees = []
+    text_inputs = []
+
     logger.info(f"{len(action_events)=}")
 
     num_events = (
@@ -195,16 +251,18 @@ def main():
         if MAX_EVENTS is not None
         else len(action_events)
     )
+
     with tqdm(
-            total=num_events,
-            desc="Preparing HTML",
-            unit="event",
-            colour="green",
-            dynamic_ncols=True,
-        ) as progress:
+        total=num_events,
+        desc="Preparing HTML",
+        unit="event",
+        colour="green",
+        dynamic_ncols=True,
+    ) as progress:
         for idx, action_event in enumerate(action_events):
             if idx == MAX_EVENTS:
                 break
+
             image = display_event(action_event)
             diff = display_event(action_event, diff=True)
             mask = action_event.screenshot.diff_mask
@@ -226,68 +284,71 @@ def main():
                 action_event_dict = scrub.scrub_dict(action_event_dict)
                 window_event_dict = scrub.scrub_dict(window_event_dict)
 
-            rows.append(
-                [
-                    row(
-                        Div(
-                            text=f"""
-                            <div class="screenshot">
-                                <img
-                                    src="{image_utf8}"
-                                    style="
-                                        aspect-ratio: {width}/{height};
-                                    "
-                                >
-                                <img
-                                    src="{diff_utf8}"
-                                    style="
-                                        aspect-ratio: {width}/{height};
-                                    "
-                                >
-                                <img
-                                    src="{mask_utf8}"
-                                    style="
-                                        aspect-ratio: {width}/{height};
-                                    "
-                                >
-                            </div>
-                            <table>
-                                {dict2html(window_event_dict , None)}
-                            </table>
-                        """,
-                        ),
-                        Div(
-                            text=f"""
-                            <table>
-                                {dict2html(action_event_dict)}
-                            </table>
-                        """
-                        ),
-                    ),
-                ]
-            )
+            with ui.column():
+                with ui.row():
+                    interactive_images.append(
+                        ui.interactive_image(
+                            source=image_utf8,
+                        ).classes("drop-shadow-md rounded")
+                    )
+            with ui.splitter(value=60) as splitter:
+                splitter.classes("w-full h-full")
+                with splitter.after:
+                    action_event_tree = create_tree(action_event_dict)
+                    action_event_trees.append(
+                        ui.tree(
+                            action_event_tree,
+                            label_key="id",
+                            on_select=lambda e: ui.notify(e.value),
+                        )
+                    )
+                    set_tree_props(action_event_trees[idx])
+                with splitter.before:
+                    ui.label("window_event_dict | action_event_dict:").style(
+                        "font-weight: bold;"
+                    )
+
+                    def on_change_closure(e, idx):
+                        return set_filter(
+                            e.value,
+                            window_event_trees[idx],
+                            action_event_trees[idx],
+                        )
+
+                    text_inputs.append(
+                        ui.input(
+                            label="search",
+                            placeholder="filter",
+                            on_change=partial(
+                                on_change_closure,
+                                idx=idx,
+                            ),
+                        )
+                    )
+                    ui.html("<br/>")
+                    window_event_tree = create_tree(window_event_dict, None)
+
+                    window_event_trees.append(
+                        ui.tree(
+                            window_event_tree,
+                            label_key="id",
+                            on_select=lambda e: ui.notify(e.value),
+                        )
+                    )
+
+                    set_tree_props(window_event_trees[idx])
 
             progress.update()
 
         progress.close()
 
-    title = f"recording-{recording.id}"
-    fname_out = f"recording-{recording.id}.html"
-    logger.info(f"{fname_out=}")
-    output_file(fname_out, title=title)
-
-    result = show(
-        layout(
-            rows,
-        )
+    ui.run(
+        reload=False,
+        title=f"OpenAdapt: recording-{recording.id}",
+        favicon="📊",
+        native=config.RUN_NATIVELY,
+        fullscreen=False,
     )
-
-    def cleanup():
-        os.remove(fname_out)
-        removed = not os.path.exists(fname_out)
-        logger.info(f"{removed=}")
-
-    Timer(1, cleanup).start()
 
 
 if __name__ == "__main__":
